Refactor analysis package.
