+++ src/main/java/lol/fairplay/ghidraapple/analysis/passes/objcclasses/OCStructureAnalyzer.kt
@@ -1,4 +1,4 @@
-package lol.fairplay.ghidraapple.analysis.objcclasses
+package lol.fairplay.ghidraapple.analysis.passes.objcclasses
 
 import ghidra.app.services.AbstractAnalyzer
 import ghidra.app.services.AnalysisPriority
